[db] Moved discs and types to separate files
