+++ libturnipripper/db_types.py
@@ -1,0 +1,3 @@
+import typing
+from typing import Iterable, Optional, ClassVar, TypeVar, Type, Union, List, Dict, Tuple, Set, Any, cast
+Json = Union[List[Any], Dict[str,Any], str, int]
